PHPStan level 7: ZERO errors — all 229 fixed
Upgraded from level 6 to level 7 and fixed all 229 errors in a
single pass across 76 files + 1 new stub file.

Level 7 adds over level 6:
  - Union type method resolution (method.notFound on union branches)
  - Stricter argument types (argument.type with union narrowing)
  - Stricter return types (return.type with union narrowing)
  - Property/offset access on union types

Fix categories (229 total):

argument.type (103):
  - strtotime() returns int|false: wrapped with (int) cast
  - json_encode() returns string|false: wrapped with (string) cast
  - file_get_contents()/session_id()/strtok(): same pattern
  - array vs list mismatches: added array_values() wrappers
  - int|string to string: added (string) cast

return.type (65):
  - Container::resolve() template annotation (fixed 35 at once)
  - list<> vs array<> return mismatches: array_values()
  - PHPDoc alignment with actual returns

method.notFound (40):
  - Created stubs/Smarty.php for CS-Cart view layer
  - Added @var annotations for dynamic property access on
    SimpleXMLElement and command dispatch patterns

offsetAccess.nonOffsetAccessible (10):
  - Precise array{} shapes for stats arrays

assign.propertyType (6):
  - Property type alignment, fopen() ?: null

foreach.nonIterable (3):
  - glob()/preg_split() ?: [] fallbacks

property.notFound (2):
  - @var object{} annotations for SimpleXMLElement children

New files:
  - stubs/Smarty.php (CS-Cart view stub for PHPStan)

phpstan: [OK] No errors (level 7, EMPTY baseline)
phpunit: 176/377/0/0

+++ addon-novoton-holidays/app/addons/novoton_holidays/src/Cron/Commands/RoomPriceCheckCommand.php
@@ -31,7 +31,7 @@ class RoomPriceCheckCommand extends AbstractCronCommand
         $nights = (int)$this->getParam('nights', 7);
         $limit = (int)$this->getParam('limit', 500);
         $country = strtoupper($this->getParam('country', ''));
-        $check_out = date('Y-m-d', strtotime($check_in . ' + ' . $nights . ' days'));
+        $check_out = date('Y-m-d', (int) strtotime($check_in . ' + ' . $nights . ' days'));
 
         $this->output("Checking hotels with active prices...");
         $this->output("Check-in: {$check_in}, Check-out: {$check_out}, Nights: {$nights}, Limit: {$limit}");
